# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Telescope REST API Swagger",
        "description": "Rest API for explorer.",
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
...
    "paths": {
        "/channels": {
            "get": {
                "tags": [
                    "Channel"
                ],
                "summary": "Query to fetch channels",
                "description": "Query to fetch channels",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Channels"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/curChannel": {
            "get": {
                "tags": [
                    "Channel"
                ],
                "summary": "get Current channel",
                "description": "get Current channel",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/curChannel"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/changeChannel/{channelName}": {
            "get": {
                "tags": [
                    "Channel"
                ],
                "summary": "Query to get Change Channel on a channel",
                "description": "Query to get Change Channel on a channel",
                "parameters": [
                    {
                        "name": "channelName",
                        "in": "path",
                        "description": "genesis_block_hash",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/changeChannel"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/block/{channelName}/{number}": {
            "get": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Query to Fetch Blocks",
                "description": "Query to Fetch Blocks",
                "parameters": [
                    {
                        "name": "channelName",
                        "in": "path",
                        "description": "genesis_block_hash",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "block number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "number"
                                },
                                "number": {
                                    "type": "string"
                                },
                                "previous_hash": {
                                    "type": "string"
                                },
                                "genesis_block_hash": {
                                    "type": "string"
                                },
                                "data_hash": {
                                    "type": "string"
                                },
                                "transactions": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "signature": {
                                                "type": "object",
                                                "properties": {
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "data": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "number"
                                                        }
                                                    }
                                                }
                                            },
                                            "payload": {
                                                "$ref": "#/definitions/blockPayload"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Channels": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "type": "array",
                "items": {
                    "type": "integer"
                }
            }
        },
        "curChannel": {
            "type": "object",
            "properties": {
                "currentChannel": {
                    "type": "string"
                }
            }
        },
        "blockPayload": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "object",
                    "properties": {
                        "channel_header": {
                            "type": "object",
                            "properties": {
                                "version": {
                                    "type": "number"
                                },
                                "type": {
                                    "type": "number"
                                },
                                "time_stamp": {
                                    "type": "string"
                                },
                                "channel_id": {
                                    "type": "string"
                                },
                                "txid": {
                                    "type": "string"
                                },
                                "epoch": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "signature_header": {
                        "type": "object",
                        "properties": {
                            "creator": {
                                "type": "object",
                                "properties": {
                                    "Mspid": {
                                        "type": "string"
                                    },
                                    "Idbytes": {
                                        "type": "string"
                                    }
                                }
                            },
                            "nonce": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "data": {
                    "type": "object"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}